Refactor: Move profile feature to a separate module
- Moved Profile API, Data, and Presentation code into dedicated modules (`feature/profile/api`, `feature/profile/data`, `feature/profile/presentation`).
- Updated package names to reflect the new module structure.
- Added necessary `build.gradle.kts` files and dependencies for the new modules.
- Made `ProfileComponent` internal within the presentation module.
- Added `ProfilePresentationModule` for Koin dependency injection within the presentation module.
