+++ feature/profile/api/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    alias(libs.plugins.holdmaster.kmplib)
+}
+
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.core.utils)
+        }
+    }
+}
+
+android {
+    namespace = "ru.bortsov.holdmaster.feature.profile.api"
+}
