+++ feature/profile/presentation/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+    alias(libs.plugins.holdmaster.kmplib)
+    alias(libs.plugins.holdmaster.compose)
+    alias(libs.plugins.holdmaster.decompose)
+    alias(libs.plugins.holdmaster.koin)
+}
+
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.feature.profile.api)
+            implementation(projects.core.utils)
+            implementation(libs.coil.compose)
+        }
+    }
+}
+
+android {
+    namespace = "ru.bortsov.holdmaster.feature.profile.presentation"
+}
