+++ feature/profile/presentation/src/commonMain/kotlin/ru/bortsov/holdmaster/feature/profile/presentation/Profile.kt
@@ -1,4 +1,4 @@
-package ru.bortsov.holdmaster.composeapp.decompose.feature.tabs.profile
+package ru.bortsov.holdmaster.feature.profile.presentation
 
 import com.arkivanov.decompose.ComponentContext
